Anchor ignore lines
The lines written to the gitignore file by write-gitignore were not
anchored. This means that a line like

  !.zsh

might cause not only ~/.zsh to be ignored, but also ~/TODO/.zsh, which
is potentially a problem.

This patch simply prepends a '/' to each entry in the gitignore file,
thereby anchoring the entry to the repository root.

Signed-off-by: martin f. krafft <madduck@madduck.net>