add sample hooks that allow cloning over extant files
Cloning a repo over an existing set of files would usually
cause a merge conflict that aborts the process and requires
manual intervention. This pair of hooks handles that case
by manually moving the extant objects out of the way,
completing the checktout, then restoring them to their
original places. The resulting state is a set of unstaged
local changes.