Fix missing blank repository name when cloning
If the repository path given to 'vcsh clone' contains a trailing slash,
the resulting cloned directory will be '.git' and vcsh will not
recognize the cloned repository by name. This change ensures that a
trailing slash will be stripped and the repository name will be
populated.

The previous faulty behavior is due to the custom implementation of the
basename command first introduced in 794f0aac.