update Arch install guide to use stable packages
Encourage the use of the stable packages per disro norms. The devel
packages can be used but they trigger full git clones in many AUR
helpers and are generally a pain if you don't specifically want them for
something. Mention the devel package but update the URLs to point to
the one that uses the normal packaging system.