Drop dependency on ronn, manpage is truly static
While the man.7 roff macro set is presentation-level and the source
is ugly, mdoc.7 (if your man page sucks, there's an excellent manual
at [1]) is mostly semantic and not as intrusive.  Readability is half
way between man.7 and markdown.

ronn is a problematic dependency as it's not widely packaged and
requires opening the rubygems can of worms.  This is mitigated by the
manpage-static branch somewhat, OTOH that approach also has its price:
various special cases in packaging code to fix up version strings,
tarball and directory names, etc.), and complicates the release process.

[1] http://mdocml.bsd.lv/man/mdoc.7.html