Make hooks work as advertised.
For vcsh commands that loop through repos, hooks were not
run correctly. Typically what would be run were:

  pre-looping-hook*
  .pre-looping-hook*
  [command run on repo r1, r2, r3 ... rN]
  post-looping-hook*
  rN.post-looping-hook

The new, more obvious behaviour is:

  pre-looping-hook*
  r1.pre-looping-hook*
  [command run on repo r1]
  r1.post-looping-hook
  [...]
  rN.pre-looping-hook*
  [command run on repo r1]
  rN.post-looping-hook
  post-looping-hook*

I was mainly concerned with the pull command, but I've made changes
in all "for VCSH_REPO_NAME" loops. This is my way of resolving
issue RichiH/vcsh#213, but as I note there are other options.