add test_setup function for setup tasks
Tests should now be written to be nilpotent; anything with lasting side
effects goes in test_setup instead of test_expect_success.  It is still
expected to succeed.  Each test_setup script will be run regardless of
which tests are being run/skipped, and their results will not be counted
in the totals.