Fix version processing on BSD systems.
BSD systems ship with the original sed; lacking a host of GNU
extensions. However both GNU sed and POSIX sed support the -E flag
for extended regular expressions which are close to the change this
is replacing.

See the descriptions of extended REs in IEEE Std 1003.2 (aka POSIX.2)
or re_format(7) on BSD systems. What matters for this is that the
+ operator works and that it and the grouping operators no longer
need to be escaped.