Alias commit to ci as well.
Mercurial, subversion and a common git alias for commit is "ci".
The PR adds that alias. Note that all of those also alias checkout
as "co" so this current case statement is a bit confusing.  That
said it doesn't make a huge amount of sense to do vcsh checkout so
it's fine if this stays if people have it in muscle memory.