Make <repo> arg to write-gitignore optional
If we're within a repository, it seems silly to expect the repository
name to be specified to write-gitignore. So this patch makes it that the
argument is optional if we're within an existing VCSH checkout. You can
still explicitly provide $2 to override the repository for which to
write the gitignore file.

Signed-off-by: martin f. krafft <madduck@madduck.net>
Signed-off-by: Caleb Maclennan <caleb@alerque.com>