add option for `push`/`pull` to run in parallel
When `-j` is given as option to `vcsh push` or `vcsh pull` it will run
its git commands in the background in parallel. This can significantly
speed up the process if more than one repo is managed by vcsh.
The `-j` option's name based on the similar option for `make`.