Load git completions if they are not already loaded
Modern bash completion uses a delayed completion loader so that
completions for commands that are never used are not loaded. This
breaks vcsh git completions if no completion for git has been
performed yet.

To resolve this, use the git completion load function to get the
completion spec for git loaded if it is not already present.