Ditch surname sort, 50% unreliable anyway
The surname sort was a nice idea, but in this case it works out very
poorly for normalization purposes:

* In many cases we only have one name, in which case it was sorting on
  email address.

* In some cases people have 3 or more name fields, it which case it was
  sorting on their middle initial.

* In some cases the entries are not even names (e.g. Fedora Release
  Engineering)

* Some languages do not follow the "first last" convention and sorting
  them based on a fixed position is just not good social etiquette.

I started trying to fix the technical problems by sorting on the *last*
field instead of blindly on the second column, but even that left so
many anomalies it was hard to find names I knew were there. I think just
a plain name sort is better in this context.