Upgrade error() + exit to fatal()
At first I was going to change `exit` to `return` in the spirit of 9357db5, but
I can't think of a good reason to just let this condition go as if it
was a success. Even if nothing needed doing in the end the fact that we
couldn't have done anything if we tried seems like it should be an error.

Co-authored-by: Dieter Plaetinck <dieter@plaetinck.be>