Use autoconf install step even on standalone builds
My previous implementation only used the configure step from autoconf.
Technically in this case it should have been enough, but it’s also kind
of an abuse of the tools, it was meant to have another stage.  This uses
the tools as they were designed, which in my history with autoconf is
always easier to maintain.  Also it gives us the option of using more
features or doing more fancy things in an install step without the
standalone versions being out of sync from system installed versions.