Make it easier to deploy a standalone script
The number of options that needed to be set to get a proper standalone
build is fairly large. Given that it was already repeated twice in
different CI contexts it occurred to me the whole batch of settings
should just be one preset so that people can build their own standalone
version easily from the source tarball if they want.