In which the need to handle $VCSH_REPO_NAME explictly becomes clear
`set -o nounset` made it clear that my flow control is really really bad
and relies heavily on $VCSH_REPO_NAME transporting information both
through its value and by being set or not. That will need to be cleaned
up.

NB: It might be an option to set it to empty as a global and leave it as
that, but that seems icky and brittle.

Signed-off-by: Richard Hartmann <richih@richih.org>