Add suport for adding main commands, plugins
vcsh is incredibly extendable and flexible however there is no
possibility to add new commands. This is were plugins come into play.
By adding plugins to $XDG_CONFIG_HOME/vcsh/plugins-enabled new commands
become available.
The plugin file name becomes the new command, the file must contain a
function with the same name as the file, this is the entry point. A help
function can also be added to get append the plugin help text to the
main help text, if help is omitted a default help message is added.