Generate checksums for release artifacts in CI
One of the major complains surrounding the recent XZ fiasco is that
auto-tools generated source files include so much obtuse code that they
are difficult to audit. Also it isn't immediately apparent what sources
they have been generated from. In our case we're generating the source
dist files in CI anyway and automatically attaching them to releases,
but GH does not make it possible to verify this. They could just as well
be reposted later by a malicious maintainer.

This is not a magic bullet to fix all that, but it should help. The CI
environment can be verified by looking at the workflow file and the
other Git sources so we're not using a modified version of autotools or
anything like that. Checksums are now being generated after making the
distribution tarballs, and *echoed to the output log* so it is possible
to verify that the files generated in CI are actually still the ones
attached to the release. The checksums file is also posted to the
release.