Include macros from autoconf-archive (pristine)
Reverting this commit should remove the macro code. The project would
still be buildable on any system with autoconf-archive, but since that
is considerably less than systems with autoconf which is our actual
target, we're vendoring a copy of them here. These files should be
bit-for-bit the same as what appears in current macro archives.